Refactor managers into corresponding service packages, add edit manager
