+++ .gitignore
@@ -1,3 +1,3 @@
-clips
+./clips
 config.yaml
 tiktok_oauth_resp.json
