+++ internal/edit/edit_manager.go
@@ -1,0 +1,10 @@
+package edit
+
+func CreateVideoWithBlurredBackground() {
+
+}
+
+
+func CreateVideoWithStackedBackground() {
+	
+}
